migrate portals to runes
